requirements.txt
deepgram_sdk==2.12.0
fastapi==0.115.5
groq==0.12.0
pydantic==2.10.1
pydantic_settings==2.6.1
python-dotenv==1.0.1
python_docx==1.1.2
setuptools==75.1.0
starlette==0.41.3
uvicorn==0.27.0
python-multipart==0.0.9